refactor: split lib into pure lib/ and I/O services/ directories; enforce directory-level coverage thresholds
