+++ biome.json
@@ -27,6 +27,6 @@
     }
   },
   "files": {
-    "includes": ["**", "!dist", "!node_modules"]
+    "includes": ["**", "!dist", "!node_modules", "!coverage"]
   }
 }
