+++ src/services/pdf-generator.ts
