+++ vitest.config.ts
@@ -6,5 +6,16 @@ export default defineConfig({
     globals: false,
     include: ['src/**/*.test.ts'],
     exclude: ['src/integration/**'],
+    coverage: {
+      provider: 'v8',
+      include: ['src/lib/**'],
+      exclude: ['src/lib/**/*.test.ts'],
+      thresholds: {
+        lines: 100,
+        functions: 100,
+        branches: 100,
+        statements: 100,
+      },
+    },
   },
 });
